+++ test/pleroma/web/activity_pub/object_validators/attachment_validator_test.exs
@@ -12,14 +12,13 @@ defmodule Pleroma.Web.ActivityPub.ObjectValidators.AttachmentValidatorTest do
 
   describe "attachments" do
     test "works with apng" do
-      attachment =
-        %{
-          "mediaType" => "image/apng",
-          "name" => "",
-          "type" => "Document",
-          "url" =>
-            "https://media.misskeyusercontent.com/io/2859c26e-cd43-4550-848b-b6243bc3fe28.apng"
-        }
+      attachment = %{
+        "mediaType" => "image/apng",
+        "name" => "",
+        "type" => "Document",
+        "url" =>
+          "https://media.misskeyusercontent.com/io/2859c26e-cd43-4550-848b-b6243bc3fe28.apng"
+      }
 
       assert {:ok, attachment} =
                AttachmentValidator.cast_and_validate(attachment)